requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pyPDF
bs4
arxiv
PyMuPDF 
wikipedia
langchain-text-splitters
lxml
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
langsmith
beautifulsoup4
streamlit
groq
langchain_groq
fastapi
uvicorn
langserve
sse_starlette
duckduckgo-search
mysql-connector-python
SQLAlchemy
validators
youtube_transcript_api
unstructured
pytube
cassio
langgraph